searcher: factor out client out of graphlbackend
This doesn't improve the client or change it at all. It just cuts and
pastes it into its own package.
